Use the BoxColliderNode class for handling actor spawn events.
