+++ Source/Engine/Level/Actors/BoxCollider.cs
@@ -5,26 +5,6 @@ using FlaxEditor.CustomEditors.Dedicated;
 
 namespace FlaxEngine
 {
-#if FLAX_EDITOR
-    /// <summary>
-    /// Dedicated custom editor for BoxCollider objects.
-    /// </summary>
-    [CustomEditor(typeof(BoxCollider)), DefaultEditor]
-    public class BoxColliderEditor : ActorEditor
-    {
-        /// <inheritdoc />
-        public override void Initialize(LayoutElementsContainer layout)
-        {
-            base.Initialize(layout);
-            layout.Space(20f);
-            var autoResizeButton = layout.Button("Resize to Fit", "Resize the box collider to fit it's parent's bounds.");
-
-            BoxCollider collider = Values[0] as BoxCollider;
-            autoResizeButton.Button.Clicked += collider.AutoResize;
-        }
-    }
-#endif
-
     partial class BoxCollider
     {
         private void BoxExcluding(Actor target, ref BoundingBox output, Actor excluded)
@@ -71,13 +51,6 @@ namespace FlaxEngine
 
             // Undo Rotation
             Orientation *= Quaternion.Invert(Orientation);
-        }
-
-        /// <inheritdoc />
-        public override void OnActorSpawned()
-        {
-            base.OnActorSpawned();
-            AutoResize();
         }
     }
 }
